project04/project04/app/requirements.txt
Flask
flask-mysql
simplejson
python-dotenv